#125 ユーザ退会 修正

+++ app/Http/Controllers/UsersController.php
@@ -47,11 +47,16 @@ class UsersController extends Controller
         return redirect()->route('user.show', $user->id);
     }
 
-    public function withdrawal()
+    public function withdrawal($id, Request $request)
     {
-        $user = Auth::user();
+        $user = User::findOrFail($id);
+        if(Auth::id() !== $user->id){
+            abort(403);
+        }
         $user->delete();
+
         Auth::logout();
+        
         return redirect("/");
     }
 }
